Hide unused lemmas.
Now searching for the regexp "^Admitted" will only find incomplete
lemmas that are depended on by the safety proof.

Incomplete lemmas meant for safety but not yet used are still
present so they can be found by SearchAbout in case they turn
out to be needed, but now have an extra argument (unused:Unused)
which is logically equivallent to false, so they can be
trivially proven and finished with Qed.

Anything only for liveness has been commented out and had
any "Admitted" lines indented to get it entirely out of the way.